requirements.txt
python-dotenv
pyTelegramBotAPI
feedparser
schedule